rename some classes and packages
