+++ ch-covidcertificate-backend-transformation/ch-covidcertificate-backend-transformation-model/src/main/java/ch/admin/bag/covidcertificate/backend/transformation/model/cert/TestCert.java
@@ -1,4 +1,4 @@
-package ch.admin.bag.covidcertificate.backend.transformation.model.pdf;
+package ch.admin.bag.covidcertificate.backend.transformation.model.cert;
 
 /**
  * This is a java implementation of the `test_entry` json spec found at
