+++ ch-covidcertificate-backend-transformation/ch-covidcertificate-backend-transformation-model/src/main/java/ch/admin/bag/covidcertificate/backend/transformation/model/pdf/BitPdfPayload.java
@@ -1,5 +1,6 @@
 package ch.admin.bag.covidcertificate.backend.transformation.model.pdf;
 
+import ch.admin.bag.covidcertificate.backend.transformation.model.cert.DecodedCert;
 import java.time.Instant;
 
 public class BitPdfPayload {
