+++ ch-covidcertificate-backend-transformation/ch-covidcertificate-backend-transformation-ws/src/main/java/ch/admin/bag/covidcertificate/backend/transformation/ws/client/BitClientMock.java
@@ -1,8 +1,8 @@
 package ch.admin.bag.covidcertificate.backend.transformation.ws.client;
 
-import ch.admin.bag.covidcertificate.backend.transformation.model.CertLightResponse;
+import ch.admin.bag.covidcertificate.backend.transformation.model.lightcert.CertLightResponse;
 import ch.admin.bag.covidcertificate.backend.transformation.model.HCertPayload;
-import ch.admin.bag.covidcertificate.backend.transformation.model.TransformPayload;
+import ch.admin.bag.covidcertificate.backend.transformation.model.lightcert.BitLightCertPayload;
 import ch.admin.bag.covidcertificate.backend.transformation.model.pdf.BitPdfPayload;
 import ch.admin.bag.covidcertificate.backend.transformation.model.pdf.PdfResponse;
 import com.fasterxml.jackson.databind.ObjectMapper;
@@ -44,7 +44,7 @@ public class BitClientMock implements BitClient {
     }
 
     @Override
-    public CertLightResponse getLightCert(TransformPayload payload) {
+    public CertLightResponse getLightCert(BitLightCertPayload payload) {
         return certLightMock;
     }
