+++ ch-covidcertificate-backend-transformation/ch-covidcertificate-backend-transformation-ws/src/main/java/ch/admin/bag/covidcertificate/backend/transformation/ws/client/PdfClient.java
@@ -4,10 +4,10 @@ import ch.admin.bag.covidcertificate.backend.transformation.model.TestResults;
 import ch.admin.bag.covidcertificate.backend.transformation.model.TestTypes;
 import ch.admin.bag.covidcertificate.backend.transformation.model.VaccinationCertificateTypes;
 import ch.admin.bag.covidcertificate.backend.transformation.model.pdf.BitPdfPayload;
-import ch.admin.bag.covidcertificate.backend.transformation.model.pdf.DecodedCert;
-import ch.admin.bag.covidcertificate.backend.transformation.model.pdf.DecodedRCert;
-import ch.admin.bag.covidcertificate.backend.transformation.model.pdf.DecodedTCert;
-import ch.admin.bag.covidcertificate.backend.transformation.model.pdf.DecodedVCert;
+import ch.admin.bag.covidcertificate.backend.transformation.model.cert.DecodedCert;
+import ch.admin.bag.covidcertificate.backend.transformation.model.cert.DecodedRCert;
+import ch.admin.bag.covidcertificate.backend.transformation.model.cert.DecodedTCert;
+import ch.admin.bag.covidcertificate.backend.transformation.model.cert.DecodedVCert;
 import ch.admin.bag.covidcertificate.backend.transformation.model.pdf.PdfResponse;
 import ch.admin.bag.covidcertificate.backend.transformation.ws.config.model.PdfConfig;
 import com.fasterxml.jackson.core.JsonProcessingException;
